OpenMP decompression
This patch adds support in zfp for OpenMP decompression.
It does that by generating an "offset table" during the compression
phase, and then using that table for performing parallel variable-rate
decoding. The patch breaks the following tests:

- testOmp: The test is broken because the behavior of chunk_size is
           changed. We need to align with the maintainers on what to
           do regarding that.

The patch has been tested locally for 1D, 2D, 3D (4D is not tested yet).
The tests will be added in a follow up once we align on the implementation.