CMake: (fix) OpenMP dependency.
In the original implementation the OpenMP libraries has been added
to the INTERFACE_LINK_LIBRARIES in the zfp-targets.cmake file using
an absolute path. This works but it is not portable. Using CMake
version 3.9 and latter the find_package(OpenMP ...) has been
significantly improved which allows for a portable solution for the
CMake config files. In addition, added ZFP_WITH_OPENMP variable to the
config file zfp-config.cmake to indicate if the zfp library has
been built with OpenMP support. If true, load the OpenMP package
using find_package(OpenMP ...).